+++ ui/src/app/services/api/magicmirror-api.service.ts
@@ -7,21 +7,21 @@ import { APIResponse, BaseAPI } from "./base-api";
 export class MagicMirrorAPI extends BaseAPI {
   public getStatus(): Promise<APIResponse> {
     console.log("Requesting MagicMirror module statuses");
-    return this.get_("mm-ctl/status");
+    return this.get_("mm/status");
   }
 
   public getUpgrade(): Promise<APIResponse> {
     console.log("Requesting MagicMirror perform upgrade");
-    return this.get_("mm-ctl/upgrade");
+    return this.get_("mm/upgrade");
   }
 
   public getInstall(): Promise<APIResponse> {
     console.log("Requesting MagicMirror perform upgrade");
-    return this.get_("mm-ctl/install");
+    return this.get_("mm/install");
   }
 
   public getRemove(): Promise<APIResponse> {
     console.log("Requesting MagicMirror perform upgrade");
-    return this.get_("mm-ctl/remove");
+    return this.get_("mm/remove");
   }
 }